update storage cards
